collect.d-backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
pandas==2.2.0
python-dotenv==1.0.0
#psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
Pillow==11.2.1
django-storages==1.14.2
boto3==1.34.34
